show hints on select
