+++ src/js/components/Sidebars.jsx
@@ -3,31 +3,36 @@ var React = require('react'),
     connect = require('react-redux').connect,
     getIn = require('../util/immutable-utils').getIn,
     ReactTooltip = require('react-tooltip'),
+    connect = require('react-redux').connect,
+    getIn = require('../util/immutable-utils').getIn,
     InspectorSidebar = require('./InspectorSidebar'),
     VisualSidebar = require('./VisualSidebar'),
     PipelinesSidebar = require('./PipelinesSidebar'),
     Toolbar = require('./Toolbar'),
-    Hints = require('./Hints'),
+    Hints = require('./walkthrough/hints'),
     Footer = require('./Footer'),
     model = require('../model');
+
 
 // Use mapDispatchToProps to force sidebar to update when the user makes any
 // change which would cause a re-render: this is clumsy but avoids forceUpdate
 function mapStateToProps(reduxState) {
+  var show = getIn(reduxState, 'walkthrough.walkthroughMode');
   return {
     // Vega "validity" is a good proxy for "has something been added or removed
     // that we need to re-render globally to account for"
-    arbitraryPropToTriggerUpdate: getIn(reduxState, 'vega.invalid')
+    arbitraryPropToTriggerUpdate: getIn(reduxState, 'vega.invalid'),
+    showHints: show
   };
 }
 
-// Splitting each sidebar into its column
 var Sidebars = React.createClass({
   propTypes: {
     arbitraryPropToTriggerUpdate: React.PropTypes.bool
   },
   render: function() {
     var pipelines = model.pipeline();
+    var showHints = this.props.showHints ? <Hints/> : '';
     return (
       <div className="sidebar-container">
         <VisualSidebar />
@@ -35,7 +40,7 @@ var Sidebars = React.createClass({
           pipelines={pipelines} />
         <PipelinesSidebar />
         <Toolbar/>
-        <Hints/>
+        {showHints}
         <Footer/>
         <ReactTooltip effect="solid"/>
       </div>
